Creada carpeta app/services para manejar la logica del programa. Accesos a la BD son realizados solo desde las clases de models.py y las funciones que lo hacian ahora realizan llamadas a las funciones de las clases, US123 arreglados y colocados en carpeta de testeo
